requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
prophet==1.1.5
pandas==2.1.4
numpy==1.26.2
firebase-admin==6.3.0
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0